fix: logicas bases de tratamento de erros para reserva e bibliotecario funcionando

+++ backend/src/services/reservaServices.js
@@ -1,5 +1,9 @@
 import { Op } from "sequelize";
 import Reserva from "../models/Reserva.js";
+import Emprestimo from "../models/Emprestimo.js";
+import Error from "../errors/reservaError.js"
+
+const { ReservaJaAssociada, ReservaNaoEncontrada, ReservaIndisponivel} = Error;
 
 class ReservaServices{
     static async findAll(){
@@ -8,25 +12,42 @@ class ReservaServices{
 
     static async findbyId(id){
         const reserva =  await Reserva.findByPk(id);
-
         if(!reserva){
-            throw new Error("Reserva não encontrada");
+            throw new ReservaNaoEncontrada();
         }
 
         return reserva
     }
 
     static async create(reserva){
-      return  await Reserva.create(
-                reserva
-            // { fields:['cpfUsuario', 'dataReserva', 'prazoReserva'] },
-        );
+        let reservaExist = await Reserva.findAll({
+            where:{
+                idLivro: reserva.idLivro ,
+                status : 'Ativa'
+            }
+        }) 
+
+        let livroEmprestado = await Emprestimo.findAll({
+            where:{
+                idLivro: reserva.idLivro ,
+                dataFim:{[Op.gte]: reserva.dataReserva}, // Empréstimo termina depois ou no início da reserva      
+            }
+        }) 
+
+        if(reservaExist.length>0){
+            throw new ReservaJaAssociada;
+        }else if(livroEmprestado.length>0){
+            throw new ReservaIndisponivel;
+        }
+
+        return await Reserva.create(reserva);
+       
     }
 
     static async update(reserva, idReserva){
         const reservaDb = await Reserva.findByPk(idReserva);
         if (!reservaDb) {
-            throw new Error("Reserva não encontrada");
+            throw new ReservaNaoEncontrada();
         }
         const reservaAtt = await reservaDb.update({ ...reserva });
         await reservaDb.save();
@@ -35,6 +56,9 @@ class ReservaServices{
 
     static async delete(id){
         const reservaDb = await Reserva.findByPk(id);
+        if (!reservaDb) {
+            throw new ReservaNaoEncontrada();
+        }
         await reservaDb.destroy();
         return reservaDb;
     }
